+++ js/utils.js
@@ -1,0 +1,16 @@
+// функция получения рандомного числа
+const getRandomPositiveInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// функция получения рандомного значения из массива
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomPositiveInteger, getRandomArrayElement};
+
+// проверка длины комментария из д/з 2 блока
+const checkStringLength = (stringLength, maxLength) => stringLength <= maxLength;
+checkStringLength(20, 40);
